Visualization for both scheduled and actual time.
The actual time a request began is now used to position the data on the
graph. The difference between scheduled time and actual time is
displayed via a smaller tail that extends back to the scheduled
execution time. The two times should ideally be identical, so the
presence of lots of visible tails is a probably a sign something is
wrong with the run.