Change from multimethod to register-handlers!
This removes the need to pass around a list of handlers, and for the
engine to know about which namespaces contain handlers. Instead, a
global registry of Transit and EDN handlers is maintained, and is used
whenever serialization occurs through the serialization namespace. The
namespaces containing the handlers must merely be loaded.